services/agents/requirements.txt
# Agents - ADK Orchestrator + Specialists
# Python 3.13

# ADK (Agent Development Kit)
# NOTA: ADK se instalará en fase posterior - no disponible en PyPI público
# google-genai-adk==1.8.0
google-generativeai==0.8.1

# Google Cloud
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-logging==3.11.3
google-auth==2.35.0

# Web framework (para endpoints HTTP de agentes)
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# HTTP client
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Google APIs (para tools)
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

